SIM-2, SIM-6: Moved all database functions to DataBaseUtils.java, deleted FileUtils.java and DataBase.java. Updated deleteRow function.
